chore: move code to a separate file
